+++ src/components/ProductTile/ProductTile.tsx
@@ -70,6 +70,13 @@ export interface ISwitcher {
     traffic: string[];
 }
 
+export interface IDefaultInfo {
+    defaultCallsValue: number;
+    defaultTrafficValue: number;
+    defaultPayment?: IServicePackPayment;
+    defaultBuyLink: string;
+}
+
 export interface IProductTileProps {
     /** Class name */
     className?: string;
@@ -126,7 +133,7 @@ export interface IProductTileProps {
     /** First params */
     firstParams: IFirstParam;
     /** Service packs */
-    servicePacks?: Array<Partial<IServicePack>>;
+    servicePacks: Array<Partial<IServicePack>>;
     /** Info - object type - return with onClickConnect, onClickBuy */
     info: {};
     /** Connect handler */
@@ -242,44 +249,61 @@ class ProductTile extends React.Component<IProductTileProps, IProductTileState> 
         usePackBuyLink: true,
     };
 
+    defaultInfo: IDefaultInfo | object;
+
     constructor(props: IProductTileProps) {
         super(props);
+
+        const {
+            usePackBuyLink,
+            servicePacks,
+            payment: { value, discount },
+            buyLink: defaultBuyLink,
+            secondParams,
+        } = props;
 
         const switcher = this.getSwitcherValues();
         const defaultCallsValue = Number(switcher.calls[props.startCallsIndex!]);
         const defaultTrafficValue = Number(switcher.traffic[props.startTrafficIndex!]);
         const currentPack = this.getCurrentPack(defaultCallsValue, defaultTrafficValue);
-        const { payment, options, buyLink, shopTag = '' } = currentPack;
+        const { payment, options, buyLink = '', shopTag = '' } = currentPack;
+        const defaultValues = {
+            defaultCallsValue,
+            defaultTrafficValue,
+            defaultPayment: payment,
+            defaultBuyLink: buyLink,
+        };
 
+        this.defaultInfo = servicePacks.length ? defaultValues : {};
         this.state = {
             switcher,
             currentPack,
             callsValue: defaultCallsValue,
             trafficValue: defaultTrafficValue,
-            price: payment && payment.value || props.payment.value,
-            discount: payment && payment.discount || props.payment.discount || '',
-            options: options || props.secondParams,
-            buyLink: props.usePackBuyLink && buyLink || this.formHashLink(props.buyLink || '', shopTag) || '',
+            price: payment && payment.value || value,
+            discount: payment && payment.discount || discount || '',
+            options: options || secondParams,
+            buyLink: usePackBuyLink && buyLink || this.formHashLink(defaultBuyLink || '', shopTag) || '',
         };
     }
 
-    handleClickConnect = (e: React.SyntheticEvent<EventTarget>) => {
-        const { onClickConnect, info } = this.props;
-        const { currentPack: { calls, traffic } } = this.state;
-        const callsValue = calls ? calls.value : 0;
-        const trafficValue = traffic ? traffic.value : 0;
-
-        onClickConnect && onClickConnect({ ...info, callsValue, trafficValue }, e);
-    }
-
-    handleClickBuy = (e: React.SyntheticEvent<EventTarget>) => {
-        const { onClickBuy, info, shopTag, payment: { unitValue, unitExtra } } = this.props;
-        const { currentPack: { calls, traffic, shopTag: packShopTag }, price, discount } = this.state;
+    getTariffInfo = () => {
+        const {
+            info,
+            shopTag,
+            payment: { unitValue, unitExtra },
+        } = this.props;
+        const {
+            currentPack: { calls, traffic, shopTag: packShopTag},
+            price,
+            discount,
+        } = this.state;
         const callsValue = calls ? calls.value : 0;
         const trafficValue = traffic ? traffic.value : 0;
         const currentShopTag = packShopTag || shopTag || '';
         const priceValue = discount ? discount : price;
-        const tariffProps = {
+
+        return {
             ...info,
             callsValue,
             trafficValue,
@@ -287,21 +311,32 @@ class ProductTile extends React.Component<IProductTileProps, IProductTileState> 
             price: priceValue,
             unitValue,
             unitExtra,
+            ...this.defaultInfo,
         };
+    }
 
-        onClickBuy && onClickBuy(tariffProps, e);
+    handleClickConnect = (e: React.SyntheticEvent<EventTarget>) => {
+        const { onClickConnect } = this.props;
+
+        onClickConnect && onClickConnect(this.getTariffInfo(), e);
+    }
+
+    handleClickBuy = (e: React.SyntheticEvent<EventTarget>) => {
+        const { onClickBuy } = this.props;
+
+        onClickBuy && onClickBuy(this.getTariffInfo(), e);
     }
 
     handleClickMore = (e: React.SyntheticEvent<EventTarget>) => {
-        const { onClickMore, info } = this.props;
+        const { onClickMore } = this.props;
 
-        onClickMore && onClickMore({ ...info }, e);
+        onClickMore && onClickMore(this.getTariffInfo(), e);
     }
 
     handleClickBubble = () => {
-        const { onClickBubble, info } = this.props;
+        const { onClickBubble } = this.props;
 
-        onClickBubble && onClickBubble({ ...info });
+        onClickBubble && onClickBubble(this.getTariffInfo());
     }
 
     handleChangeCalls = (_e: React.SyntheticEvent<EventTarget>, value: string): boolean => {